+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Add Hash command
+- Add `airdrop:hash` command.
+- Add GitHub Actions driver.
 
 ## 0.1.0 - 2021-05-16